classification/requirement.txt
pillow
numpy
plotly
tensorflow
matplotlib
argparse
pyfiglet